Fix `:transport_options` allowing non-typical values
Work was previously done to support non-typical values in the `:transport_options`
for ThousandIsland with https://github.com/mtrudel/bandit/commit/43d49c8817b2c2f71bf4149036f5680dae36b02d.

However, if using SSL and including multiple key/value pairs with a non-typical,
such as `:inet6`, it would still error because of an attempt to `Keyword.delete/2`
out the `:otp_app` key.

This fixes that with `Enum.reject/2` to get the same behavior and still allow
the list as expected by `:ssl` and `:gen_tcp`