Start using `:proc_lib.set_label/1` if available
Starting in OTP 27.0, `:proc_lib.set_label/1` allows setting labels on
processes whether or not they have a registered name. These labels are
shown in Observer and in the future may be shown in other tools.

Using such labels has the potential to make a process tree more
comprehensible. This commit just serves as an example for getting
started.