Support using zstd compression if available (#514)
* Use zstd compression if available

* Fix applying zstd options

* Prefer zstd when available

Originally we would default to the first encoding found. Now we apply
our own weighting.

A typical browser header for accept encoding is `gzip, deflate, br,
zstd` for reference.

* Fix failing tests

* Pass zstd_options directly to :zstd

The types for zstd were not exported, thus this makes it a bit easier
to not have to document them all.

It comes with a performance benefit of also not having to keep
fetching from a keyword list and letting users even define their own
dictionaries.

* Make zstd encoding opt in

* Revert content encoding priority setting

This is a seperate discussion that should not block on supporting a
new compression format.

This returns the content encoding negotiation to picking the first one
the client sends that we support.

* Fix compiler warnings on erlang/otp <28